add tabs and usePatient hook
