+++ client/packages/system/src/Patient/EditPatientModal/EditPatientModal.tsx
@@ -1,56 +1,30 @@
-import React, { useEffect, useMemo } from 'react';
+import React, { useState } from 'react';
 import {
   DetailContainer,
-  useConfirmationModal,
   Box,
   useTranslation,
-  UpdatePatientInput,
-  BasicSpinner,
-  DocumentRegistryCategoryNode,
-  UserPermission,
   DialogButton,
   useDialog,
-  useNotification,
   SaveIcon,
   LoadingButton,
+  DetailSection,
+  TabContext,
+  Tab,
+  TabList,
+  BasicSpinner,
   useAuthContext,
+  UserPermission,
 } from '@openmsupply-client/common';
-import { usePatient } from '../api';
-import {
-  JsonFormData,
-  FormInputData,
-  SavedDocument,
-  SchemaData,
-  useDocumentRegistry,
-  useJsonFormsHandler,
-  usePatientStore,
-  PatientSchema,
-  SaveDocumentMutation,
-} from '@openmsupply-client/programs';
 
-import defaultPatientSchema from './DefaultEditPatientSchema.json';
-import defaultPatientUISchema from './DefaultEditPatientUISchema.json';
+import { usePatientEditForm } from './usePatientEditForm';
+import { PatientPanel } from '../CreatePatientModal/PatientPanel';
+import { useInsuranceProviders } from '../apiModern/hooks/useInsuranceProviders';
+import { PatientInsuranceTab } from './PatientInsuranceTab';
 
-const DEFAULT_SCHEMA: SchemaData = {
-  formSchemaId: undefined,
-  jsonSchema: defaultPatientSchema,
-  uiSchema: defaultPatientUISchema,
-};
-
-const useUpsertProgramPatient = (): SaveDocumentMutation => {
-  const { mutateAsync: updateProgramPatient } =
-    usePatient.document.updateProgramPatient();
-
-  return async (jsonData: unknown, schemaId: string, parent?: string) => {
-    const result = await updateProgramPatient({
-      data: jsonData,
-      parent: parent ? parent : '',
-      schemaId,
-    });
-    if (!result.document) throw Error('Inserted document not set!');
-    return result.document;
-  };
-};
+enum Tabs {
+  Patient = 'Patient',
+  Insurance = 'Insurance',
+}
 
 export const EditPatientModal = ({
   isOpen,
@@ -62,180 +36,38 @@ export const EditPatientModal = ({
   onClose: (patientId?: string) => void;
 }) => {
   const t = useTranslation();
+  const [currentTab, setCurrentTab] = useState(Tabs.Patient);
+
+  const { JsonForm, save, isLoading, isSaving, isDirty, validationError } =
+    usePatientEditForm(patientId, onClose);
   const { userHasPermission } = useAuthContext();
-
-  const {
-    documentName,
-    setDocumentName,
-    createNewPatient,
-    // setCreateNewPatient,
-  } = usePatientStore();
-
-  const { mutateAsync: updatePatient } = usePatient.document.update();
-
-  const { error } = useNotification();
 
   const { Modal } = useDialog({
     onClose,
     isOpen,
   });
 
-  const { data: currentPatient, isLoading: isCurrentPatientLoading } =
-    usePatient.document.get(patientId);
-  const { data: patientRegistries, isLoading: isPatientRegistryLoading } =
-    useDocumentRegistry.get.documentRegistries({
-      filter: {
-        category: { equalTo: DocumentRegistryCategoryNode.Patient },
-      },
-    });
-
-  const isLoading = isCurrentPatientLoading || isPatientRegistryLoading;
-
-  const patientRegistry = patientRegistries?.nodes[0];
-
-  // we have to memo the data to avoid an infinite render loop
-  const inputData = useMemo<FormInputData | undefined>(() => {
-    if (!!currentPatient && !currentPatient.document) {
-      // The loaded patient doesn't has a document. Use the information we got
-      // (from the name table).
-      return {
-        schema: patientRegistry ?? DEFAULT_SCHEMA,
-        data: {
-          id: currentPatient.id,
-          code: currentPatient.code,
-          code2: currentPatient.code2 ?? undefined,
-          firstName: currentPatient.firstName ?? undefined,
-          lastName: currentPatient.lastName ?? undefined,
-          gender: currentPatient.gender ?? undefined,
-          dateOfBirth: currentPatient.dateOfBirth ?? undefined,
-          dateOfDeath: currentPatient.dateOfDeath ?? undefined,
-          isDeceased: currentPatient.isDeceased ?? undefined,
-          phone: currentPatient.phone ?? undefined,
-          address1: currentPatient.address1 ?? undefined,
-          nextOfKin:
-            currentPatient.nextOfKinId || currentPatient.nextOfKinName
-              ? {
-                  id: currentPatient.nextOfKinId ?? undefined,
-                  name: currentPatient.nextOfKinName ?? undefined,
-                }
-              : undefined,
-          extension: {},
-        },
-        isCreating: false,
-      };
-    } else if (currentPatient?.document) {
-      // Take the data from the document
-      return {
-        schema: patientRegistry ?? DEFAULT_SCHEMA,
-        data: currentPatient.documentDraft,
-        isCreating: false,
-      };
-    }
-  }, [createNewPatient, currentPatient, patientRegistry]);
-
-  const isCreatingPatient = false;
-
-  const handleProgramPatientSave = useUpsertProgramPatient();
-
-  const handlePatientSave = async (input: UpdatePatientInput) => {
-    await updatePatient(input);
-  };
-
-  const accessor: JsonFormData<SavedDocument | void> = patientRegistry
-    ? {
-        loadedData: inputData?.data,
-        isLoading: false,
-        error: undefined,
-        isCreating: isCreatingPatient,
-        schema: patientRegistry,
-        save: async (data: unknown) => {
-          await handleProgramPatientSave(
-            data,
-            patientRegistry.formSchemaId,
-            currentPatient?.document?.id
-          );
-        },
-      }
-    : {
-        loadedData: inputData?.data,
-        isLoading: false,
-        error: undefined,
-        isCreating: isCreatingPatient,
-        schema: DEFAULT_SCHEMA,
-        save: async (data: unknown) => {
-          const patientData = data as PatientSchema;
-          const newData = Object.fromEntries(
-            Object.entries(data ?? {}).filter(
-              ([key]) =>
-                key !== 'dateOfBirthIsEstimated' &&
-                key !== 'nextOfKin' &&
-                key !== 'extension'
-            )
-          );
-          // map nextOfKin object to individual fields
-          const input = {
-            ...newData,
-            nextOfKinId: patientData?.nextOfKin?.id,
-            nextOfKinName: patientData?.nextOfKin?.name,
-          };
-          await handlePatientSave(input as UpdatePatientInput);
-        },
-      };
-
-  const { JsonForm, saveData, isSaving, isDirty, validationError } =
-    useJsonFormsHandler(
-      {
-        documentName: createNewPatient ? undefined : documentName,
-        patientId: patientId,
-      },
-      accessor
-    );
-
-  const handleSave = async () => {
-    try {
-      const savedDocument = await saveData();
-
-      if (savedDocument) {
-        setDocumentName(savedDocument.name);
-      }
-      onClose();
-    } catch (e) {
-      const errorSnack = error((e as Error).message);
-      errorSnack();
-    }
-  };
-
-  const save = useConfirmationModal({
-    title: t('heading.are-you-sure'),
-    message: t('messages.confirm-save-generic'),
-    onConfirm: handleSave,
-  });
-
-  useEffect(() => {
-    if (!documentName && currentPatient) {
-      setDocumentName(currentPatient?.document?.name);
-    }
-  }, [currentPatient, documentName, setDocumentName]);
+  const {
+    query: { data: insuranceProvidersData },
+  } = useInsuranceProviders();
 
   if (isLoading) return <BasicSpinner />;
-
-  if (isLoading) {
-    return null;
-  }
 
   return (
     <Modal
       title=""
       width={950}
+      height={650}
       okButton={
         <LoadingButton
           color="secondary"
           disabled={
             !isDirty ||
+            isSaving ||
             !!validationError ||
             !userHasPermission(UserPermission.PatientMutate)
           }
-          isLoading={isSaving}
+          isLoading={isLoading}
           onClick={() => save()}
           label={t('button.save')}
           startIcon={<SaveIcon />}
@@ -251,9 +83,38 @@ export const EditPatientModal = ({
       }
       slideAnimation={false}
     >
-      <DetailContainer>
-        <Box>{JsonForm}</Box>
-      </DetailContainer>
+      <TabContext value={currentTab}>
+        <Box display="flex" flexDirection="column" alignItems="center" gap={2}>
+          <DetailSection title="">
+            <Box flex={1}>
+              <TabList
+                value={currentTab}
+                centered
+                onChange={(_, v) => setCurrentTab(v)}
+              >
+                <Tab
+                  value={Tabs.Patient}
+                  label={t('label.patient-details')}
+                  tabIndex={-1}
+                />
+                {insuranceProvidersData.length > 0 && (
+                  <Tab
+                    value={Tabs.Insurance}
+                    label={t('label.insurance')}
+                    tabIndex={-1}
+                  />
+                )}
+              </TabList>
+            </Box>
+          </DetailSection>
+        </Box>
+        <DetailContainer>
+          <PatientPanel value={Tabs.Patient}>{JsonForm}</PatientPanel>
+          <PatientPanel value={Tabs.Insurance}>
+            <PatientInsuranceTab />
+          </PatientPanel>
+        </DetailContainer>
+      </TabContext>
     </Modal>
   );
 };
